+++ src/Components/EnergyCalculator/Steps/GasProvider.tsx
@@ -61,7 +61,6 @@ export default function GasProvider() {
     };
 
     const updatedFormData: FormData = { ...formData, energyCalculator: updatedEnergyData };
-    setFormData(updatedFormData);
     updateScreen(updatedFormData);
     nextStep();
   };
