+++ src/Components/EnergyCalculator/Steps/Utilities.tsx
@@ -72,7 +72,6 @@ const Utilities = () => {
 
     const updatedEnergyCalculatorData = { ...formData.energyCalculator, ...rhfData.energyCalculator };
     const updatedFormData = { ...formData, energyCalculator: updatedEnergyCalculatorData };
-    setFormData(updatedFormData);
     updateScreen(updatedFormData);
     nextStep();
   };
