+++ src/Components/Steps/Referrer.tsx
@@ -86,7 +86,6 @@ export default function ReferralSourceStep() {
   const formSubmitHandler = async ({ referralSource, otherReferrer }: FormSchema) => {
     const source = otherReferrer !== '' ? otherReferrer : referralSource;
     const updatedFormData = { ...formData, referralSource: source };
-    setFormData(updatedFormData);
     updateScreen(updatedFormData);
     nextStep();
   };
