+++ src/Components/Steps/SelectLanguage.tsx
@@ -78,7 +78,7 @@ const SelectLanguagePage = () => {
     }
 
     updateWhiteLabelAndNavigate(stateCodes[0], `/${stateCodes[0]}/step-2${queryString}`);
-    setFormData({ ...formData, whiteLabel: stateCodes[0] });
+    setFormData({ ...formData, whiteLabel: stateCodes[0] }); // TODO: Is this necessary? We have a separate state for white label, and it is not populated if the page refreshes
     // wait for the new config to be loaded
     const interval = setInterval(() => {
       if (!configLoading) {
